Implement a preliminary draft for Game
First change is a new constructor for Game which takes two Players. As
Players have not been added, there is a stub implementation for them in
this commit. The same is true for Move and History.

The new makeMove method can be used to safely apply a Move to the board.
It currently does some basic verification on Move and saves the Move in
a History instance for further use.